requirements.txt
anyio==3.6.2
bcrypt==4.0.1
bson==0.5.10
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.1
pydantic==1.10.5
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
